backend/app/requirements.txt
aerich
aiohttp
aiomysql
aiosqlite
amqp
anyio
async-generator
async-timeout
asyncclick
asyncpg
atlastk
attrs
bcrypt
billiard
cachetools
celery
certifi
cffi
chardet
ciso8601
click
cryptography
cssselect
cssutils
databases
dnspython
email-validator
emails
fastapi
flower
future
h11
h2
hpack
hstspreload
httpcore
httptools
httpx
httpx-oauth
humanize
hyperframe
idna
importlib-metadata
inflection
iso8601
Jinja2
kombu
lxml
makefun
Mako
MarkupSafe
more-itertools
multidict
packaging
passlib
pluggy
premailer
prometheus-client
py
pycparser
pydantic
PyJWT
PyMySQL
pyparsing
PyPika
pytest
python-dateutil
python-editor
python-multipart
pytz
redis
requests
rfc3986
six
sniffio
SQLAlchemy
starlette
tornado
tortoise-orm
typing-extensions
urllib3
uvicorn
uvloop
vine
wcwidth
websockets
yarl
zipp
python-decouple